--- assets/js/base/components/cart-checkout/form/style.scss
@@ -1,4 +1,0 @@
-.wc-block-components-checkout-form {
-	margin: 0;
-	max-width: 100%;
-}
