+++ assets/js/blocks/cart-checkout/checkout/form/login-prompt.js
@@ -1,0 +1,33 @@
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+import { CHECKOUT_SHOW_LOGIN_REMINDER } from '@woocommerce/block-settings';
+import { useCheckoutContext } from '@woocommerce/base-context';
+
+/**
+ * Internal dependencies
+ */
+import { LOGIN_TO_CHECKOUT_URL } from '../utils';
+
+const LoginPrompt = () => {
+	const { customerId } = useCheckoutContext();
+
+	if ( ! CHECKOUT_SHOW_LOGIN_REMINDER || customerId ) {
+		return null;
+	}
+
+	return (
+		<>
+			{ __(
+				'Already have an account? ',
+				'woo-gutenberg-products-block'
+			) }
+			<a href={ LOGIN_TO_CHECKOUT_URL }>
+				{ __( 'Log in.', 'woo-gutenberg-products-block' ) }
+			</a>
+		</>
+	);
+};
+
+export default LoginPrompt;
