+++ assets/js/blocks/cart-checkout/checkout/style.scss
@@ -2,16 +2,6 @@
 	top: -96px;
 }
 
-.wc-block-checkout__use-address-for-billing {
-	margin-top: em($gap-large);
-}
-
-.wc-block-checkout__shipping-option {
-	.wc-block-components-shipping-rates-control__package:not(:first-of-type) {
-		margin-top: $gap-larger;
-	}
-}
-
 .wc-block-checkout__sidebar {
 	.wc-block-components-product-name {
 		color: inherit;
@@ -62,7 +52,7 @@
 		@include force-content();
 		width: 150px;
 	}
-	.wc-block-components-checkout-form {
+	.wc-block-checkout__form {
 		.wc-block-components-checkout-step__title {
 			@include placeholder();
 			@include force-content();
@@ -156,58 +146,8 @@
 	}
 }
 
-.is-small,
-.is-medium,
-.is-large {
-	.wc-block-checkout__billing-fields,
-	.wc-block-checkout__shipping-fields {
-		.wc-block-components-address-form {
-			margin-left: #{-$gap-small / 2};
-			margin-right: #{-$gap-small / 2};
-
-			&::after {
-				content: "";
-				clear: both;
-				display: block;
-			}
-
-			.wc-block-components-text-input,
-			.wc-block-components-country-input,
-			.wc-block-components-state-input {
-				float: left;
-				margin-left: #{$gap-small / 2};
-				margin-right: #{$gap-small / 2};
-				position: relative;
-				width: calc(50% - #{$gap-small});
-
-				&:nth-of-type(2),
-				&:first-of-type {
-					margin-top: 0;
-				}
-			}
-
-			.wc-block-components-address-form__company,
-			.wc-block-components-address-form__address_1,
-			.wc-block-components-address-form__address_2 {
-				width: calc(100% - #{$gap-small});
-			}
-
-			.wc-block-components-checkbox {
-				clear: both;
-			}
-
-		}
-	}
-}
-
 .is-large {
 	.wc-block-checkout__actions {
 		padding-right: 36px;
-	}
-
-	.wc-block-checkout__shipping-option {
-		.wc-block-components-radio-control__input {
-			margin-left: -8px;
-		}
 	}
 }
