+++ assets/js/blocks/cart-checkout/checkout/utils.js
@@ -1,0 +1,3 @@
+export const LOGIN_TO_CHECKOUT_URL = `/wp-login.php?redirect_to=${ encodeURIComponent(
+	window.location.href
+) }`;
