+++ docs/theming/README.md
@@ -72,3 +72,4 @@ WooCommerce Blocks avoids using legacy unprefixed classes as much as possible. H
 -   [Product grid blocks style update in 2.7.0](./product-grid-270.md)
 -   [Class names update in 2.8.0](./class-names-update-280.md)
 -   [Class names update in 3.3.0](./class-names-update-330.md)
+-   [Class names update in 3.4.0](./class-names-update-340.md)
