+++ docs/theming/class-names-update-340.md
@@ -1,0 +1,14 @@
+# Class names update in 3.4.0
+
+## Replaced classes
+
+<table>
+	<tr>
+		<th>Removed</th>
+		<th>New class</th>
+	</tr>
+	<tr>
+		<td><code>wc-block-components-checkout-form</code></td>
+		<td><code>wc-block-checkout__form</code></td>
+	</tr>
+</table>
