+++ src/BlockTypes/Checkout.php
@@ -208,7 +208,7 @@ class Checkout extends AbstractBlock {
 				<div class="wc-block-components-main wc-block-checkout__main">
 					<div class="wc-block-components-express-payment wc-block-components-express-payment--checkout"></div>
 					<div class="wc-block-components-express-payment-continue-rule wc-block-components-express-payment-continue-rule--checkout"><span></span></div>
-					<form class="wc-block-components-checkout-form">
+					<form class="wc-block-checkout__form">
 						<fieldset class="wc-block-checkout__contact-fields wc-block-components-checkout-step">
 							<div class="wc-block-components-checkout-step__heading">
 								<div class="wc-block-components-checkout-step__title"></div>
